+++ chocolatey-visualstudio.extension/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 1.9.0
 - Added custom handling for '--installCatalogUri' to better support installations from layout without '--noWeb' ([GH-65](https://github.com/jberezanski/ChocolateyPackages/pull/65)).
 - New switch for Add-VisualStudioWorkload: -IncludeOptionalComponentsByDefault ([GH-92](https://github.com/jberezanski/ChocolateyPackages/pull/92)).
+- Fixed support for Visual Studio 2019 16.9 installer, which is no longer based on Electron.
 
 ## Version 1.8.1